Create a new tracing backend based on the Z3 backend.
The new backend uses the same logic as the Z3 backend for everything except for
exploring other paths.  It currently checks for consistency of the path
condition but that is probably not necessary.  After that check, it records each
path constraint on a log file.  This is a very inefficient way of saving the
path condition, we may optimize it later on.

Also, this backend copies most of the code for the tracing backend, we should
refactor the common code and make the two backends differ only on the main entry
point `_sym_push_path_constraint`.  We can explore other ways to build symbolic
expressions later on for performance.